Introducing Decicion tree, Random Forest and Extreme Random Forest
Wrapper_RF is created to prepare the data for the trees. IT is based on
the SVM wrapper. But here we do not need to search for subfeatures.
Threfore the whole brute fore and the greedy forward search are deleted.
The Wrapper just creates a train and test feature set. VAlidation
feature set is not needed as we do not itterate the Trees.

Tree Methods are:
Decicion Tree
Random Forest
Extreme Random Forest

The Classifier_routines now have a function called
Classifier_random_forest

Here we use the train data to train and the test set to validate the
perofmance.
This is done over all patients with leave one patient out for the test
set.