Use different probability threshold to determine class
!That change happend already two commits earlier. Just for commenting on
it

Classifier_routine
* We calculate the probabilities for each class. Then we determine a
threshold other than the default 0.5 to determine a class.
It compares all classes other than AS and chooses that class if the
probability is higher than a threshold.

Next step would be to do this with a small gridsearch from 0.2 to 0.5
for each patent a choose the class decition with the highest performance