Annotation Changer & Normalisation per session
---------------------------------
Loading_5min_mat_files_cECG
* Added a annotationchanger to the loading process. Now we can change
all CLasses 6 (IS) To the following class. Also the class 5(NA) is
changed if it is sourrounded by the same class and last less then x
amount of epochs. It is then changed to the sourrounding class

* Normalisation is now done in the Loading_5min_... file instead of the
wrapper file because we also now normalize on each session and then
merge the sessions to ive them to the wrapper