Multiple classifier in a wrapper
#-----------------------------
The wraper now collects all variable information from the user and calls
loading data and a leave one out cross validation (LOOCV). Afterwards it
merges the predictions of all the classifers used and forms one ultimate
prediction which is that evaluated b Kappa and F1

#-------------------------------------
LOOCV
* In the leave one out cross validation the test and training sets are
generated and gien to the Random Forrest classifier function.
It returns the test results and the probabilities for the classes of the
Random forrest.
This can be called many times with differnt input parameters
#----------------------------------
* Loading 5min_...
Here the data is just loaded, normalized. They can be polynominalized
and/or change the annotations if asked from the wrapper.

#-----------------------------------
Classifier routines
* the entropy criterion, number of estimators and leave depth are now
moved to the wrapper to make it possible to change them per run.